pyproject.toml
[project]
name = "mcpserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.6.0",
    "langgraph-sdk>=0.1.66",
    "langgraph-checkpoint>=2.0.23",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langsmith>=0.4.5",
    "httpx>=0.25.0",
    "sse-starlette>=2.1.0,<2.2.0",
    "uvloop>=0.18.0",
    "httptools>=0.5.0",
    "jsonschema-rs>=0.20.0",
    "structlog>=24.1.0",
    "cloudpickle>=3.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "anyio>=4.7.0",
    "fastmcp>=2.10.4",
    "grpcio>=1.73.1",
    "pydantic>=2.11.7",
    "qdrant-client>=1.14.3",
    "openai>=1.95.1",
    "uvicorn>=0.35.0",
    "fastembed>=0.7.3",
    "asyncpg>=0.30.0",
    "flake8-import-order>=0.19.2",
    "flake8-builtins>=3.1.0",
    "pep8-naming>=0.15.1",
    "flake8-bugbear>=25.10.21",
    "flake8-commas>=4.0.0",
    "flake8-variables-names>=0.0.6",
    "flake8-functions>=0.0.8",
    "flake8-class-attributes-order>=0.3.0",
    "flake8-expression-complexity>=0.0.11",
    "flake8-cognitive-complexity>=0.1.0",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
    "pytest-asyncio>=1.2.0"
]

[tool.uv]
# всегда подтягивать dev-зависимости
default-groups = ["dev"]


[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",      # pycodestyle
  "F",      # pyflakes
  "I",      # isort
  "UP",     # pyupgrade
  "D",      # pydocstyle (включая D100/D101/D102/D103/D104/D107 и D401)
  "T201",   # запрет print()
  "G",      # logging (G004: запрет f-string в logger.*)
  "RUF100", # контроль noqa (лишние/неиспользуемые)
]

ignore = [
  "E501",  # длина строки (контролируем line-length/format)
  "D417",  # не требуем документировать все аргументы (у тебя Args в description=)
  # Конфликты по оформлению docstring у классов:
  # выбираем D211/D212-стиль (без пустой строки перед class docstring, summary в первой строке)
  "D203",  # 1 blank line required before class docstring (конфликтует с D211)
  "D213",  # Multi-line docstring summary should start at the second line (конфликтует с D212)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D", "UP"]
"src/**/__init__.py" = ["D104"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
lines-after-imports = 2

# Если используешь ruff format:
# ruff format будет приводить код к единому виду, а ruff check будет проверять правила.
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

 

[tool.mypy]
plugins = ["pydantic.mypy"] # Включает плагин Pydantic для корректной типизации моделей и автогенерации сигнатур конструкторов
ignore_missing_imports = true # глобально игнорирует отсутствие type stubs для сторонних библиотек,
follow_imports = "silent" # Не выводить предупреждения о типах в импортируемых модулях (ускоряет проверку и уменьшает шум)
warn_redundant_casts = true # Предупреждать о ненужных операциях приведения типа (например, cast, которые не влияют на вывод типов)
warn_unused_ignores = true # Предупреждать, если используется # type: ignore, которые фактически не нужны (чистота кода)
disallow_any_generics = true # Запретить обобщенные типы с неявным any (например, List без указания типа элементов)
check_untyped_defs = true # Проверять функции без аннотаций типов (строже, помогает найти ошибки в неаннотированном коде)
no_implicit_reexport = true # Не экспортировать типы импорты из __init__.py автоматически (контролирует импортируемые типы)
disallow_untyped_defs = true # Требовать аннотаций типов у всех функций и методов (повышает строгость проверки)


[tool.pydantic-mypy]
init_forbid_extra = true # Запретить передачу в конструктор моделей Pydantic полей, не описанных явно в модели
init_typed = true # Генерировать типизированные сигнатуры __init__ на основе типов атрибутов модели (типизация конструктора)
warn_required_dynamic_aliases = true # Предупреждать, если используются алиасы или динамические имена полей без явной типизации
